requirements.txt
faiss-cpu
sentence-transformers
numpy
pydantic
pyyaml
tqdm
streamlit
duckdb
astor
openai
python-dotenv
groq
